third_party/blink/web_tests/http/tests/devtools/network/network-xhr-replay-expected.txt

Tests XHR replaying. Bug 95187


Running: testGetStaticAsync
Dumping request: 
    url: http://user:password@127.0.0.1:8000/devtools/network/resources/empty.html
    requestMethod: GET
    test request header value: headerValueA

Running: testGetStaticSync
Dumping request: 
    url: http://user:password@127.0.0.1:8000/devtools/network/resources/empty.html
    requestMethod: GET
    test request header value: headerValueB

Running: testGetCachedAsync
Dumping request: 
    url: http://user:password@127.0.0.1:8000/devtools/network/resources/random-cached.php
    requestMethod: GET
    test request header value: headerValueC

Running: testGetCachedSync
Dumping request: 
    url: http://user:password@127.0.0.1:8000/devtools/network/resources/random-cached.php
    requestMethod: GET
    test request header value: headerValueD

Running: testGetRandomAsync
Dumping request: 
    url: http://user:password@127.0.0.1:8000/devtools/network/resources/random.php
    requestMethod: GET
    test request header value: headerValueE

Running: testGetRandomSync
Dumping request: 
    url: http://user:password@127.0.0.1:8000/devtools/network/resources/random.php
    requestMethod: GET
    test request header value: headerValueF

Running: testPostAsync
Dumping request: 
    url: http://user:password@127.0.0.1:8000/devtools/network/resources/random.php
    requestFormData: payload
    requestMethod: POST
    test request header value: headerValueG

Running: testPostSync
Dumping request: 
    url: http://user:password@127.0.0.1:8000/devtools/network/resources/random.php
    requestFormData: payload
    requestMethod: POST
    test request header value: headerValueH
